requirements/requirements.txt
Django==2.2
django-ckeditor==6.0.0
django-js-asset==1.2.2
djongo==1.3.4
psycopg2==2.8.6
psycopg2-binary==2.8.6
pymongo==3.11.3
pytz==2021.1
sqlparse==0.2.4
